fix: watch dataset store variable in analysis page
the SubHeader component is responsible for rendering and mutating the
selected variable of interest. Currently it only sets the variable on
the dataset store. The TimeSeriesPlot is watching the dataset store for
changes but in the analysis page has been special cased to use traces
pulled into the analysis store (with additional statistics, timeseries,
etc.)

a future refactor should more coherently unify these so we don't need to
special case these kinds of mutations

resolves #162